@@ -1,20 +1,19 @@
 {
   "$schema": "../node_modules/ng-packagr/package.schema.json",
   "name": "@braune-digital/angular-query-filter",
-  "version": "2.0.8",
+  "version": "11.0.0",
   "description": "Components to use with the query-filter-bundle.",
   "author": "Braune Digital GmbH <mail@braune-digital.com>",
   "license": "GNU",
   "repository": "git@github.com:braune-digital/angular-query-filter.git",
   "peerDependencies": {
-    "@angular/core": ">=9.0.0",
-    "@angular/common": ">=9.0.0",
-    "@angular/forms": ">=9.0.0",
-    "ngx-bootstrap": ">=5.5.0",
+    "@angular/core": ">=11.0.0",
+    "@angular/common": ">=11.0.0",
+    "@angular/forms": ">=11.0.0",
+    "ngx-bootstrap": ">=6.2.0",
     "date-fns": ">=1.30",
-    "@braune-digital/form-utilities": ">=2.0.0",
-    "rxjs": ">=5.5",
-    "@ngx-translate/core": ">=12.0.0",
+    "@braune-digital/form-utilities": ">=11.0.0",
+    "@ngx-translate/core": ">=13.0.0",
     "ts-md5": ">=1.2.0",
     "qs": ">=6.9.0"
   },
