@@ -15,51 +15,51 @@
     "postinstall": "ngcc"
   },
   "dependencies": {
-    "@angular/animations": "10.0.3",
-    "@angular/common": "10.0.3",
-    "@angular/compiler": "10.0.3",
-    "@angular/core": "10.0.3",
-    "@angular/forms": "10.0.3",
-    "@angular/platform-browser": "10.0.3",
-    "@angular/platform-browser-dynamic": "10.0.3",
-    "@angular/router": "10.0.3",
+    "@angular/animations": "11.2.13",
+    "@angular/common": "11.2.13",
+    "@angular/compiler": "11.2.13",
+    "@angular/core": "11.2.13",
+    "@angular/forms": "11.2.13",
+    "@angular/platform-browser": "11.2.13",
+    "@angular/platform-browser-dynamic": "11.2.13",
+    "@angular/router": "11.2.13",
     "core-js": "^2.4.1",
     "date-fns": "^1.30.1",
-    "ngx-bootstrap": "^5.6.1",
-    "qs": "^6.9.4",
+    "ngx-bootstrap": "^6.2.0",
+    "qs": "^6.10.1",
     "rxjs": "^6.5.4",
     "ts-md5": "^1.2.4",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.2",
-    "@angular/cli": "10.0.2",
-    "@angular/compiler-cli": "10.0.3",
-    "@angular/language-service": "10.0.3",
+    "@angular-devkit/build-angular": "~0.1102.12",
+    "@angular/cli": "11.2.12",
+    "@angular/compiler-cli": "11.2.13",
+    "@angular/language-service": "11.2.13",
     "@braune-digital/form-utilities": "^2.0.2",
-    "@ng-select/ng-select": "^2.20.5",
-    "@ngx-translate/core": "^11.0.1",
-    "@types/jasmine": "~2.8.3",
+    "@ng-select/ng-select": "^6.1.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^10.0.0",
-    "ng2-ckeditor": "^1.2.7",
-    "ngx-quill": "^4.8.0",
+    "ng-packagr": "^11.2.4",
+    "ng2-ckeditor": "^1.3.3",
+    "ngx-quill": "^13.4.0",
     "ngx-textarea-autosize": "^2.0.0",
     "protractor": "~7.0.0",
     "quill": "^1.3.6",
     "rimraf": "^2.6.2",
     "ts-node": "~4.1.0",
     "tslint": "~6.1.0",
-    "typescript": "^3.9.6"
+    "typescript": "^4.1.5"
   }
 }
